successfully connect java to mysql db
