+++ market_mind_web/.gitignore
@@ -31,4 +31,3 @@ build/
 
 ### VS Code ###
 .vscode/
-*.vscode/
