+++ market_mind_web/src/main/java/com/market_mind/market_mind_web/MarketMindWebApplication.java
@@ -1,4 +1,4 @@
-package com.market_mind_web;
+package com.market_mind.market_mind_web;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
