+++ market_mind_web/src/main/resources/application.properties
@@ -1,3 +1,4 @@
-spring.datasource.url=jdbc:mysql://${DB_HOST}:3306/${DB_NAME}
-spring.datasource.username=${DB_USERNAME}
+spring.datasource.url=jdbc:mysql://${DB_HOST}:3306/${DB_NAME}?useSSL=false&serverTimezone=UTC
+spring.datasource.username=${DB_USER}
 spring.datasource.password=${DB_PASSWORD}
+spring.jpa.hibernate.ddl-auto=none
